Move timer server to common.
