+++ common/src/lib.rs
@@ -1,5 +1,5 @@
 //
-// ReZe.Rs - Router Daemon
+// ReZe.Rs - Common
 //   Copyright (C) 2018-2020 Toshiaki Takada
 //
 // Common library
@@ -8,5 +8,6 @@
 pub mod consts;
 pub mod error;
 pub mod event;
+pub mod timer;
 pub mod uds_server;
 pub mod acl;
